requirements.txt
flask==3.1.3
flask-socketio==5.5.1
joblib==1.5.3
pandas==3.0.2
xgboost==3.2.0
smbus2==0.4.3
paramiko==3.5.1
python-dotenv==1.0.1
